# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Yandex Zen Article Generator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Type of article to generate
 */
export type GenerateArticleRequestArticleType =
  (typeof GenerateArticleRequestArticleType)[keyof typeof GenerateArticleRequestArticleType];

export const GenerateArticleRequestArticleType = {
  review: "review",
  guide: "guide",
  listicle: "listicle",
  news: "news",
  opinion: "opinion",
} as const;

/**
 * Writing tone
 */
export type GenerateArticleRequestTone =
  (typeof GenerateArticleRequestTone)[keyof typeof GenerateArticleRequestTone];

export const GenerateArticleRequestTone = {
  formal: "formal",
  casual: "casual",
  expert: "expert",
  friendly: "friendly",
} as const;

/**
 * Approximate article length
 */
export type GenerateArticleRequestLength =
  (typeof GenerateArticleRequestLength)[keyof typeof GenerateArticleRequestLength];

export const GenerateArticleRequestLength = {
  short: "short",
  medium: "medium",
  long: "long",
} as const;

export interface GenerateArticleRequest {
  /** The main topic or title of the article */
  topic: string;
  /** Category of the article (e.g. lifestyle, tech, products, travel) */
  category: string;
  /** SEO keywords to include */
  keywords?: string[];
  /** Type of article to generate */
  articleType: GenerateArticleRequestArticleType;
  /** Product URL to include as affiliate link (optional) */
  productUrl?: string;
  /** Target audience description */
  targetAudience?: string;
  /** Writing tone */
  tone?: GenerateArticleRequestTone;
  /** Approximate article length */
  length?: GenerateArticleRequestLength;
}

export interface GeneratedArticle {
  id: number;
  title: string;
  /** HTML content of the article */
  content: string;
  /** Short excerpt/lead paragraph */
  excerpt: string;
  seoTitle: string;
  metaDescription: string;
  keywords: string[];
  category: string;
  articleType: string;
  /** Suggested image descriptions for the article */
  imageDescriptions: string[];
  wordCount: number;
  /** Estimated reading time in minutes */
  readingTime: number;
  createdAt: string;
}

export interface SavedArticle {
  id: number;
  title: string;
  content: string;
  excerpt: string;
  seoTitle: string;
  metaDescription: string;
  keywords: string[];
  category: string;
  articleType: string;
  imageDescriptions: string[];
  wordCount: number;
  readingTime: number;
  topic: string;
  productUrl?: string;
  createdAt: string;
}

export interface ArticleListResponse {
  articles: SavedArticle[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type ArticleStatsByCategory = { [key: string]: number };

export type ArticleStatsByType = { [key: string]: number };

export interface ArticleStats {
  totalArticles: number;
  byCategory: ArticleStatsByCategory;
  byType: ArticleStatsByType;
  avgWordCount: number;
  avgReadingTime: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type ListArticlesParams = {
  page?: number;
  limit?: number;
};