# lib/api-zod/src/generated/types/articleListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Yandex Zen Article Generator
 * OpenAPI spec version: 0.1.0
 */
import type { SavedArticle } from "./savedArticle";

export interface ArticleListResponse {
  articles: SavedArticle[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}