# lib/api-zod/src/generated/types/articleStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Yandex Zen Article Generator
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleStatsByCategory } from "./articleStatsByCategory";
import type { ArticleStatsByType } from "./articleStatsByType";

export interface ArticleStats {
  totalArticles: number;
  byCategory: ArticleStatsByCategory;
  byType: ArticleStatsByType;
  avgWordCount: number;
  avgReadingTime: number;
}