# lib/api-zod/src/generated/types/generateArticleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Yandex Zen Article Generator
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateArticleRequestArticleType } from "./generateArticleRequestArticleType";
import type { GenerateArticleRequestLength } from "./generateArticleRequestLength";
import type { GenerateArticleRequestTone } from "./generateArticleRequestTone";

export interface GenerateArticleRequest {
  /** The main topic or title of the article */
  topic: string;
  /** Category of the article (e.g. lifestyle, tech, products, travel) */
  category: string;
  /** SEO keywords to include */
  keywords?: string[];
  /** Type of article to generate */
  articleType: GenerateArticleRequestArticleType;
  /** Product URL to include as affiliate link (optional) */
  productUrl?: string;
  /** Target audience description */
  targetAudience?: string;
  /** Writing tone */
  tone?: GenerateArticleRequestTone;
  /** Approximate article length */
  length?: GenerateArticleRequestLength;
}