# lib/api-zod/src/generated/types/generatedArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Yandex Zen Article Generator
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedArticle {
  id: number;
  title: string;
  /** HTML content of the article */
  content: string;
  /** Short excerpt/lead paragraph */
  excerpt: string;
  seoTitle: string;
  metaDescription: string;
  keywords: string[];
  category: string;
  articleType: string;
  /** Suggested image descriptions for the article */
  imageDescriptions: string[];
  wordCount: number;
  /** Estimated reading time in minutes */
  readingTime: number;
  createdAt: Date;
}