# lib/api-zod/src/generated/types/savedArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Yandex Zen Article Generator
 * OpenAPI spec version: 0.1.0
 */

export interface SavedArticle {
  id: number;
  title: string;
  content: string;
  excerpt: string;
  seoTitle: string;
  metaDescription: string;
  keywords: string[];
  category: string;
  articleType: string;
  imageDescriptions: string[];
  wordCount: number;
  readingTime: number;
  topic: string;
  productUrl?: string;
  createdAt: Date;
}